[Chat-frontend]: Code Cleanup and Refactoring
- Added values.js file for refactoring string and numeral constants
- Fixed Height bug in tab items
- Created index.js for components for easy access to components
