+++ src/front_end/chat/src/App.js
@@ -3,9 +3,9 @@ import React, { useState, useEffect } from "react";
 import { Container } from '@material-ui/core'
 import { BrowserRouter, Switch, Route } from "react-router-dom";
 
-import Navbar from "./components/Navbar/Navbar";
-import Home from "./components/Home/Home";
-import SplashScreen from './components/SplashScreen/SplashScreen'
+import { values } from "./constants";
+
+import { Navbar, Home, SplashScreen } from './components'
 
 import './App.css'
 
@@ -14,11 +14,11 @@ const App = () => {
     // Loading state variable
     const [isLoading, setIsLoading] = useState(true);
 
-    // Waiting for 3 seconds before routing to main
+    // Waiting for animation timeout seconds before routing to main
     useEffect(() => {
         setTimeout(() => {
             setIsLoading(false);
-        }, 4700);
+        }, values.SPLASH_SCREEN_ANIMATION_TIMEOUT);
     }, []);
