+++ src/front_end/chat/src/api/index.js
@@ -2,16 +2,6 @@ import axios from 'axios'
 
 const API = axios.create({ baseURL: "http://4c94-35-238-238-160.ngrok.io" });
 
-export const fetchPosts = () => API.get('/posts');
-export const createPost = (newPost) => API.post('/posts', newPost);
-export const updatePost = (id, updatedPost) => API.patch(`/posts/${id}`, updatedPost);
-export const deletePost = (id) => API.delete(`/posts/${id}`);
-export const likePost = (id) => API.patch(`/posts/${id}/likePost`);
-
-export const signin = (formData) => API.post(`/users/signin`, formData);
-export const signup = (formData) => API.post(`/users/signup`, formData);
-
-// Used for MUKALMA
 export const sendMessage = (message) => API.post('/reply', message)
 export const fetchSource = () => API.get('/source')
 export const fetchTopics = () => API.get('/topics')
