+++ src/front_end/chat/src/components/InformationRetrieval/styles.css
@@ -2,7 +2,7 @@
     position: relative;
     border: 1px solid gainsboro;
     width: 100%;
-    height: 450px;
+    height: 100%;
     display: flex;
     flex-direction: column;
     justify-content: space-evenly;
