+++ src/front_end/chat/src/components/TabContainer/TabContainer.css
