+++ src/front_end/chat/src/components/index.js
@@ -1,0 +1,9 @@
+export { default as Chatbox } from './Chatbox/Chatbox'
+export { default as Home } from './Home/Home'
+export { default as InformationRetrieval } from './InformationRetrieval/InformationRetrieval'
+export { default as Navbar } from './Navbar/Navbar'
+export { default as Responses } from './Responses/Responses'
+export { default as Source } from './Source/Source'
+export { default as SplashScreen } from './SplashScreen/SplashScreen'
+export { default as TabContainer } from './TabContainer/TabContainer'
+export { default as Topic } from './Topic/Topic'
