+++ src/front_end/chat/src/constants/index.js
@@ -1,2 +1,3 @@
 export { default as images } from './images'
 export { default as audio } from './audio'
+export { default as values } from './values'
